quando si registra associa un carrello correttamente

+++ home-health-care/src/main/java/it/univpm/hhc/model/dao/SubDao.java
@@ -18,7 +18,7 @@ public interface SubDao {
 	
 	List <Sub> findbySub(Long id);
 
-	Sub create(String name, Long price, int discount);
+	Sub create(String name, Double price, int discount);
 	
 	Sub update(Sub sub);
